requirements.txt
asn1crypto==0.24.0
captcha==0.2.4
cffi==1.11.5
cryptography==2.3.1
diff-match-patch==20121119
Django==2.0.8
django-cors-middleware==1.3.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-recaptcha==1.4.0
django-reversion==3.0.0
django-simple-captcha==0.5.9
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
jdcal==1.4
msgpack==0.5.6
odfpy==1.3.6
openpyxl==2.5.8
Pillow==5.0.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
requests
django-debug-toolbar
django-rest-swagger
python-weixin
markdown
django-filter
django-guardian
coreapi
djangorestframework-recursive
social-auth-app-django
django-redis
django-rest-multiple-models
 rest-framework-generic-relations